refactor o reflect change for more generic database
