+++ apps/src/main/java/org/jgrasstools/nww/layers/defaults/spatialite/RasterizedSpatialiteLasLayer.java
@@ -41,12 +41,12 @@ import org.geotools.referencing.crs.DefaultGeographicCRS;
 import org.jgrasstools.dbs.compat.ASpatialDb;
 import org.jgrasstools.dbs.compat.GeometryColumn;
 import org.jgrasstools.dbs.spatialite.SpatialiteGeometryColumns;
-import org.jgrasstools.gears.io.las.spatialite.LasCell;
-import org.jgrasstools.gears.io.las.spatialite.LasCellsTable;
-import org.jgrasstools.gears.io.las.spatialite.LasLevel;
-import org.jgrasstools.gears.io.las.spatialite.LasLevelsTable;
-import org.jgrasstools.gears.io.las.spatialite.LasSource;
-import org.jgrasstools.gears.io.las.spatialite.LasSourcesTable;
+import org.jgrasstools.gears.io.las.databases.LasCell;
+import org.jgrasstools.gears.io.las.databases.LasCellsTable;
+import org.jgrasstools.gears.io.las.databases.LasLevel;
+import org.jgrasstools.gears.io.las.databases.LasLevelsTable;
+import org.jgrasstools.gears.io.las.databases.LasSource;
+import org.jgrasstools.gears.io.las.databases.LasSourcesTable;
 import org.jgrasstools.gears.utils.CrsUtilities;
 import org.jgrasstools.gears.utils.TransformationUtils;
 import org.jgrasstools.gears.utils.colors.ColorInterpolator;
