+++ jgrassgears/src/main/java/org/jgrasstools/gears/io/las/databases/LasCellsTable.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.jgrasstools.gears.io.las.spatialite;
+package org.jgrasstools.gears.io.las.databases;
 
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
